added wrap as parameter to Accessor type
Wrap intends to denote that while some accessors are transparent, some only
permit access with checks (Maybe) or permit access to more than one value of the
sub type (List).

In order for accessors through these kinds of wrapping layers to make sense, we
are forced to return the result of the sub Accessor getter wrapped into this
layer of wrapping. Wrap lets us keep track of which wrappings have not
disappeared.